[cmake_frontend.py] compilation db frontend
